refactor: 패키지 레이어에 맞게 분류
