+++ src/main/java/subway/station/presentation/StationController.java
@@ -1,8 +1,9 @@
-package subway.station;
+package subway.station.presentation;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.application.StationService;
 
 import java.net.URI;
 import java.util.List;
