+++ src/test/java/subway/section/application/SectionServiceTest.java
@@ -1,4 +1,4 @@
-package subway.section;
+package subway.section.application;
 
 import static org.junit.jupiter.api.Assertions.*;
