+++ src/test/java/subway/section/domain/SectionTest.java
@@ -1,9 +1,10 @@
-package subway.section;
+package subway.section.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import subway.section.domain.Section;
 
 import static org.assertj.core.api.Assertions.assertThat;
