+++ src/test/java/subway/section/domain/SectionsTest.java
@@ -1,9 +1,12 @@
-package subway.section;
+package subway.section.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import subway.section.domain.Section;
+import subway.section.domain.Sections;
+import subway.section.presentation.SectionRequest;
 
 import java.util.Arrays;
 import java.util.Collections;
